% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{check_library_sheet}
\alias{check_library_sheet}
\title{Check library sheet for consistency}
\usage{
check_library_sheet(sheet, n_samples)
}
\arguments{
\item{sheet}{data.frame, library sheet to be tested}

\item{n_samples}{number of samples expected}
}
\value{
TRUE If library sheet passes all test, otherwise a string describing
the first test that failed.
}
\description{
Check library sheet for consistency
}
