% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comuneqaid.R
\name{detect_chemistry}
\alias{detect_chemistry}
\title{Infer chemistry for sequencing run}
\usage{
detect_chemistry(bcl_folder)
}
\arguments{
\item{bcl_folder}{path to bcl input folder}
}
\value{
string, version of chemistry used
}
\description{
Infer chemistry for sequencing run
}
