% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{detect_samplesheet_version}
\alias{detect_samplesheet_version}
\title{Detect sample sheet version}
\usage{
detect_samplesheet_version(sheet_names)
}
\arguments{
\item{sheet_names}{vector of the names of sheets in excel file}
}
\value{
string, sample sheet version
}
\description{
Detect sample sheet version
}
