% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_flowcell_id}
\alias{extract_flowcell_id}
\title{Extract flowcell ID from folder name}
\usage{
extract_flowcell_id(x)
}
\arguments{
\item{x}{string, BCL folder name}
}
\value{
string containing the flowcell ID
}
\description{
Extract flowcell ID from folder name
}
\examples{
extract_flowcell_id("900101_A00001_0001_ABCDEFGHI1")
}
