% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_pin}
\alias{extract_pin}
\title{Extract BCL pin from folder name}
\usage{
extract_pin(x)
}
\arguments{
\item{x}{string, BCL folder name}
}
\value{
string containing the four digits in the pin
}
\description{
Extract BCL pin from folder name
}
\examples{
extract_pin("900101_A00001_0001_ABCDEFGHI1")
}
