% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_sequencing_date}
\alias{extract_sequencing_date}
\title{Extract sequencing date from folder name}
\usage{
extract_sequencing_date(x)
}
\arguments{
\item{x}{string, BCL folder name}
}
\value{
string containing the sequencing date
}
\description{
Extract sequencing date from folder name
}
\examples{
extract_sequencing_date("900101_A00001_0001_ABCDEFGHI1")
}
