% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{fix_library_sheet}
\alias{fix_library_sheet}
\title{Fix typical issues in sample sheets}
\usage{
fix_library_sheet(sheet)
}
\arguments{
\item{sheet}{data.frame read by \link{read_input_files}.}
}
\value{
sample sheet with simple common errors fixed
}
\description{
Fix typical issues in sample sheets
}
