% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comuneqaid.R
\name{from_pin_to_bcl}
\alias{from_pin_to_bcl}
\title{Infer BCL folder from PIN}
\usage{
from_pin_to_bcl(sheet, bcl_path)
}
\arguments{
\item{sheet}{data.frame generated by \link{read_input_files}}

\item{bcl_path}{path to bcl input folder}
}
\value{
the data frame in 'sheet' where 10x_pin and hto_pin has been replaced
with 10x_bcl and hto_bcl
}
\description{
Infer BCL folder from PIN
}
