% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_resources.R
\name{get_resources}
\alias{get_resources}
\title{Get location specific resources}
\usage{
get_resources(location, species)
}
\arguments{
\item{location}{string, name of computing location.}

\item{species}{string, name of species investigated.}
}
\value{
list of resource locations
}
\description{
Get location specific resources
}
