% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{guess_species}
\alias{guess_species}
\title{Guess Species}
\usage{
guess_species(x)
}
\arguments{
\item{x}{character, species}
}
\value{
string, standardized species name
}
\description{
Guess Species
}
\examples{
guess_species("mouse")
}
