% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comuneqaid.R
\name{hto_features}
\alias{hto_features}
\title{Generate HTO feature tables}
\usage{
hto_features(sheet)
}
\arguments{
\item{sheet}{data.frame generated by \link{read_input_files}}
}
\value{
list of data.frames
}
\description{
Generate HTO feature tables
}
