% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comuneqaid.R
\name{infer_and_check_config}
\alias{infer_and_check_config}
\title{Infer config parameters and sanity check them}
\usage{
infer_and_check_config(config, sheet)
}
\arguments{
\item{config}{list of config settings}

\item{sheet}{data.frame generated by \link{read_input_files}}

\item{chemistry}{string, chemistry version}
}
\value{
list of updated config settings
}
\description{
Infer config parameters and sanity check them
}
