% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comuneqaid.R
\name{make_pool_table}
\alias{make_pool_table}
\title{Create pool table}
\usage{
make_pool_table(sheet, split_by = c("i7_index", "lane"))
}
\arguments{
\item{sheet}{data.frame generated by \link{read_input_files}}

\item{split_by}{character vector of column names to split the sheet by}
}
\value{
data.frame containing information about the 10x pools
}
\description{
Create pool table
}
