% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comuneqaid.R
\name{make_sample_sheet}
\alias{make_sample_sheet}
\title{Create sample sheet}
\usage{
make_sample_sheet(i7, hto = NULL, version, flowcell_id)
}
\arguments{
\item{i7}{vector of strings containing 10x index set names}

\item{hto}{vector of strings containing HTO index names}

\item{version}{string, chemistry version}

\item{flowcell_id}{string, containing 10 digit flowcell id}
}
\value{
string containing sample sheet
}
\description{
Create sample sheet
}
