% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{notin}
\alias{notin}
\alias{\%!in\%}
\title{Inverse Value Matching}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.}

\item{table}{vector or \code{NULL}: the values to be matched against.}
}
\value{
A logical vector, indicating if a match was nor located for each
element of \code{x}: thus the values are \code{TRUE} or \code{FALSE} and
never \code{NA}.
}
\description{
Inverse Value Matching
}
\examples{
c("a", "b", "c") \%!in\% c("c", "d")
}
\seealso{
\link[base]{match}
}
