% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{parse_app_data}
\alias{parse_app_data}
\title{Parse files from COMUNEQAID app}
\usage{
parse_app_data(samp, nume, sele, text)
}
\arguments{
\item{samp}{string, path to samp.csv file}

\item{nume}{string, path to nume.csv file}

\item{sele}{string, path to sele.csv file}
}
\value{
data.frame with info necessary for running COMUNEQAID
}
\description{
Parse files from COMUNEQAID app
}
