% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{parse_config}
\alias{parse_config}
\title{Parse a config file}
\usage{
parse_config(file)
}
\arguments{
\item{file}{string, path to config file with sequencing information}
}
\value{
list of data.frames with info tables
}
\description{
Parse a config file
}
