% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{parse_library_sheet_v2}
\alias{parse_library_sheet_v2}
\title{Parse a library sheet v2 file}
\usage{
parse_library_sheet_v2(file)
}
\arguments{
\item{file}{string, path to excel file with sequencing information}
}
\value{
list of data.frames with info tables
}
\description{
Parse a library sheet v2 file
}
