% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comuneqaid.R
\name{prepare_comuneqaid}
\alias{prepare_comuneqaid}
\title{Prepare for COMUNEQAID pipeline}
\usage{
prepare_comuneqaid(
  comuneqaid_id,
  input_files,
  config = list(),
  lib_sheet_path = default_lib_sheet_path,
  qc_path = default_qc_path,
  bcl_path = default_bcl_path,
  fastq_path = default_fastq_path,
  out_path = default_out_path,
  log_path = default_log_path,
  config_path = default_config_path
)
}
\arguments{
\item{comuneqaid_id}{required string, project ID}

\item{input_files}{required vector of input file names usually xlsx file,
see details.}

\item{config}{list of user defined config settings, can be empty list.
If set, overrides all other settings including paths and comuneqaid id
given to this function.}

\item{lib_sheet_path}{optional string, path to library sheets folder}

\item{qc_path}{optional string, path to qc output folder}

\item{bcl_path}{optional string, path to bcl input folder}

\item{fastq_path}{optional string, path to fastq output folder}

\item{out_path}{optional string, path to results output folder}

\item{log_path}{optional string, path to logs output folder}

\item{config_path}{optional string, path to config file output folder}
}
\value{
invisibly returns updated config list
}
\description{
Prepare for COMUNEQAID pipeline
}
\details{
Input files can be either xlsx file(s) or text-files from the
COMUNEQAID app. Using multiple xlsx files is untested and may break in new
and exciting ways.
}
