% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_comuneqaid.R
\name{prepare_folders}
\alias{prepare_folders}
\title{Prepare folder structure for COMUNEQAID}
\usage{
prepare_folders(
  qc_path,
  bcl_path,
  bcl_folders,
  comuneqaid_id,
  fastq_path,
  out_path,
  log_path,
  config_path
)
}
\arguments{
\item{qc_path}{string, path to where qc output should go}

\item{bcl_path}{string, path to bcl input folder}

\item{bcl_folders}{vector of strings, bcl folders}

\item{comuneqaid_id}{string, comuneqaid run ID}

\item{fastq_path}{string, path to bcl output folder}

\item{out_path}{string, path to results output folder}

\item{log_path}{string, path to logs output folder}

\item{config_path}{string, path to config file output folder}
}
\value{
invisibly returns NULL
}
\description{
Prepare folder structure for COMUNEQAID
}
