% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{process_seq_info}
\alias{process_seq_info}
\title{Process sequencing info sheet}
\usage{
process_seq_info(seq_info, sample_col, output_col)
}
\arguments{
\item{seq_info}{data.frame, sequencing info}

\item{sample_col}{string, name of column with comma separated sample names}

\item{output_col}{string, name of column with bcl folder}
}
\value{
data.frame mapping samples to BCL folders
}
\description{
Process sequencing info sheet
}
