% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{read_input_files}
\alias{read_input_files}
\title{Figure out the type of input and use the correct input}
\usage{
read_input_files(files, lib_sheet_path)
}
\arguments{
\item{files}{vector of files to be processed}

\item{lib_sheet_path}{string, path to library sheets folder}
}
\value{
data.frame with sample information
}
\description{
Figure out the type of input and use the correct input
}
