% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_library_sheet.R
\name{read_multiple_sheets}
\alias{read_multiple_sheets}
\title{Read multiple library sheets at once}
\usage{
read_multiple_sheets(
  library_sheets,
  reader = parse_library_sheet_v1,
  lib_sheet_path = default_lib_sheet_path
)
}
\arguments{
\item{library_sheets}{vector of library sheet filenames}

\item{reader}{function used to read library_sheets}

\item{lib_sheet_path}{character path to library sheets, if unspecified
uses `default_lib_sheet_path`}
}
\value{
single merges library sheet
}
\description{
Read multiple library sheets at once
}
\details{
If the library sheets are located in different folders, the full
path can be given in `library_sheets` and `lib_sheet_path` set to an empty
string.
TODO: Probably doesn't handle re-sequencing. Number of samples will be wrong.
}
\examples{
\dontrun{
read_multiple_sheets(vector_of_sheet_paths)
}
}
