% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{write_log}
\alias{write_log}
\title{Write message to log}
\usage{
write_log(
  message = NULL,
  section = FALSE,
  title = FALSE,
  log_number = NULL,
  total_logs = NULL,
  horizontal_line = FALSE,
  width = 80L
)
}
\arguments{
\item{message}{string, text to be printed}

\item{section}{logical, format message as section header}

\item{title}{logical, format message as title}

\item{log_number}{optional integer, denoting current log number. Ignored
unless title = TRUE}

\item{total_logs}{optional integer, denoting total number of logs. Ignored
unless title = TRUE}

\item{horizontal_line}{logical, ignore message and print horizontal line}

\item{width}{integer, line width in characters}
}
\value{
a formatted message is printed to standard out, \code{NULL} is
invisibly returned
}
\description{
Write message to log
}
\examples{
write_log(c("foo", "bar"))
write_log(c("foo", "bar"), section = TRUE)
write_log(c("foo", "bar"), title = TRUE)
write_log(c("foo", "bar"), title = TRUE, log_number = 4, total_logs = 6)
write_log(horizontal_line = TRUE)
write_log("a very long line of characters", width = 25, title = TRUE)
}
